+++ src/app/apply/apply.tsx
@@ -1,14 +1,14 @@
 import Head from 'next/head';
-import Header from '@/components/Header';
-import ImageHighlight from '@/components/ImageHighlight';
-import Container from '@/components/Container';
-import LinkButton from '@/components/LinkButton';
-import AnimatedStagger, { StaggerItem } from '@/components/AnimatedStagger';
+import Header from '@/components/common/Header';
+import ImageHighlight from '@/components/common/ImageHighlight';
+import Container from '@/components/common/Container';
+import LinkButton from '@/components/common/LinkButton';
+import AnimatedStagger, { StaggerItem } from '@/components/common/AnimatedStagger';
 import styles from '@/app/apply/apply.module.scss';
 import classNames from 'classnames';
-import QuestionCell from '@/components/QuestionCell';
+import QuestionCell from '@/components/apply/QuestionCell';
 import Link from 'next/link';
-import Hover from '@/components/Hover';
+import Hover from '@/components/common/Hover';
 import { Application } from '@/types/contentful';
 import { documentToHtmlString } from '@contentful/rich-text-html-renderer';
 import { IoMdArrowDropright as ArrowIcon } from 'react-icons/io';
