+++ src/components/about/ExecBoard.module.scss
