+++ src/components/common/ImageHighlight.module.scss
