+++ src/components/home/AnimatedRole.module.scss
