+++ src/components/home/Card.module.scss
